Add KeyboardSpan component
In light of keeping some of the same functionality of the original AIM
experience that I don't fully remember, this gets the first letter of
the command, and checks when the alt key is pressed by itself. After that,
it will listen for the next key press and perform the action passed in to
the listener prop.

- Other style changes and updates to the textarea to keep state in tact