Update chatLoader, add room check when posting messages
This updates the chatRouterLoader to add a cookie set function to track
which room the user is in. This value is checked when posting messages to
prevent users that are logged in from posting messages to rooms they aren't
currently in, i.e. via Copy -> As fetch and changing the room search param.

- Update the chatLoader to return an object and type useLoaderData consumers