Add rate limiter
This forks the oak_rate_limiter package and attaches it to a new /debug/test
route. test.sh sends a bunch of requests and it seems to work.

Deno doesn't like the rate limiter, but I think it's due to some version issue
with oak. All efforts of clearing cache have failed (sans nuking everything).